platform-api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
pytest>=8.0.0
allure-pytest>=2.13.5
